# doc/api.yaml
openapi: 3.0.0
paths:
...
  /artist:
    get:
      operationId: ArtistController_findAll
      summary: Get all artists
      description: Gets all artists
      parameters: []
      responses:
        '200':
          description: Artists have been successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArtistResponseDto'
      tags: &ref_2
        - artist
      security: &ref_3
        - bearer: []
    post:
      operationId: ArtistController_create
      summary: Add new artist
      description: Add new artist
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArtistDto'
      responses:
        '201':
          description: The artist has been successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponseDto'
        '400':
          description: Request body does not contain required fields
      tags: *ref_2
      security: *ref_3
  /artist/{id}:
    get:
      operationId: ArtistController_findOne
      summary: Get single artist by id
      description: Get single artist by id
      parameters:
        - name: id
          required: true
          in: path
          description: artist id (uuid v4)
          example: 7b06da94-6e57-4641-8baa-6f6d975c70aa
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: The artist has been successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponseDto'
        '400':
          description: Provided id of artist is invalid (not uuid)
        '404':
          description: Artist with provided id was not found
      tags: *ref_2
      security: *ref_3
    put:
      operationId: ArtistController_updateInfo
      summary: Update artist information
      description: Update artist information by UUID
      parameters:
        - name: id
          required: true
          in: path
          description: artist id (uuid v4)
          example: 70083ecc-00e2-45fd-b7fb-1a7b58eab45f
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtistInfoDto'
      responses:
        '200':
          description: Artist's info has been successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponseDto'
        '400':
          description: Provided id of artist is invalid (not uuid)
        '404':
          description: Artist with provided id was not found
      tags: *ref_2
      security: *ref_3
    delete:
      operationId: ArtistController_remove
      summary: Delete artist
      description: Delete artist from library
      parameters:
        - name: id
          required: true
          in: path
          description: artist id (uuid v4)
          example: 25f82d5b-e19f-405d-bafa-191930b0577b
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Artist has been successfully deleted
        '400':
          description: Provided id of artist is invalid (not uuid)
        '404':
          description: Artist with provided id was not found
      tags: *ref_2
      security: *ref_3
  /album:
    get:
      operationId: AlbumController_findAll
      summary: Get albums list
      description: Gets all library albums list
      parameters: []
      responses:
        '200':
          description: Albums have been successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlbumResponseDto'
      tags: &ref_4
        - album
      security: &ref_5
        - bearer: []
    post:
      operationId: AlbumController_create
      summary: Add new album
      description: Add new album information
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlbumDto'
      responses:
        '201':
          description: The album has been successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponseDto'
        '400':
          description: >-
            Request body does not contain required fields or artistId in DTO is
            invalid or points to non-existing entity
      tags: *ref_4
      security: *ref_5
  /album/{id}:
    get:
      operationId: AlbumController_findOne
      summary: Get single album by id
      description: Gets single album by id
      parameters:
        - name: id
          required: true
          in: path
          description: album id (uuid v4)
          example: 0a91b6a8-d1af-4556-80e0-f482e33232a0
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: The album has been successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponseDto'
        '400':
          description: Provided id of album is invalid (not uuid)
        '404':
          description: Album with provided id was not found
      tags: *ref_4
      security: *ref_5
    put:
      operationId: AlbumController_updateInfo
      summary: Update album information
      description: Update library album information by UUID
      parameters:
        - name: id
          required: true
          in: path
          description: album id (uuid v4)
          example: ab00a15e-86c9-4ed1-84e0-3234eb315b2b
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlbumInfoDto'
      responses:
        '200':
          description: Album's info has been successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumResponseDto'
        '400':
          description: >-
            Provided id of album is invalid (not uuid) or artistId in DTO is
            invalid or points to non-existing entity
        '404':
          description: Album with provided id was not found
      tags: *ref_4
      security: *ref_5
    delete:
      operationId: AlbumController_remove
      summary: Delete album
      description: Delete album from library
      parameters:
        - name: id
          required: true
          in: path
          description: album id (uuid v4)
          example: a868adda-61a6-4d4f-9ba9-43629fa73147
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Album has been successfully deleted
        '400':
          description: Provided id of album is invalid (not uuid)
        '404':
          description: Album with provided id was not found
      tags: *ref_4
      security: *ref_5
...
  /favs/album/{id}:
    post:
      operationId: FavsController_addAlbum
      summary: Add album to the favorites
      description: Add album to the favorites
      parameters:
        - name: id
          required: true
          in: path
          description: album id (uuid v4)
          example: 21d1e499-b101-43bf-aa68-42bc782ca0b2
          schema:
            format: uuid
            type: string
      responses:
        '201':
          description: The album has been successfully added to favorites
        '400':
          description: Provided id of album is invalid (not uuid)
        '422':
          description: Album with provided id was not found
      tags: *ref_8
      security: *ref_9
    delete:
      operationId: FavsController_removeAlbum
      summary: Delete album from favorites
      description: Delete album from favorites
      parameters:
        - name: id
          required: true
          in: path
          description: album id (uuid v4)
          example: 652025d4-37cf-49f5-87a1-88e95774007d
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Album has been successfully deleted from the favorites
        '400':
          description: Provided id of album is invalid (not uuid)
        '404':
          description: Album with provided id was not found in the favorites
      tags: *ref_8
      security: *ref_9
  /favs/artist/{id}:
    post:
      operationId: FavsController_addArtist
      summary: Add artist to the favorites
      description: Add artist to the favorites
      parameters:
        - name: id
          required: true
          in: path
          description: artist id (uuid v4)
          example: 8abbc4b2-a0df-4eba-8384-f5077d0d6ed5
          schema:
            format: uuid
            type: string
      responses:
        '201':
          description: The artist has been successfully added to favorites
        '400':
          description: Provided id of artist is invalid (not uuid)
        '422':
          description: Artist with provided id was not found
      tags: *ref_8
      security: *ref_9
    delete:
      operationId: FavsController_removeArtist
      summary: Delete artist from favorites
      description: Delete artist from favorites
      parameters:
        - name: id
          required: true
          in: path
          description: artist id (uuid v4)
          example: 8f62dd6d-db0c-4a93-85d0-850b2a55b8fd
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Artist has been successfully deleted from the favorites
        '400':
          description: Provided id of artist is invalid (not uuid)
        '404':
          description: Artist with provided id was not found in the favorites
      tags: *ref_8
      security: *ref_9
...
info:
  title: Home Library Service
  description: 'Home Library Service: RS School NodeJS 2023 Q2 - Weeks 6-9 Task'
  version: 0.0.1
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    ArtistResponseDto:
      type: object
      properties:
        id:
          type: string
          description: artist ID (uuid v4)
          example: 3606733e-3bea-489c-a045-afa6e63ac0b3
        name:
          type: string
          description: artist name
          example: Björk
        grammy:
          type: boolean
          description: Does artist have Grammy awards?
          example: true
      required:
        - id
        - name
        - grammy
    CreateArtistDto:
      type: object
      properties:
        name:
          type: string
          description: artist name
          example: Moby
        grammy:
          type: boolean
          description: Does artist have Grammy awards?
          example: true
      required:
        - name
        - grammy
    UpdateArtistInfoDto:
      type: object
      properties:
        name:
          type: string
          description: artist name
          example: Harry Styles
        grammy:
          type: boolean
          description: Does artist have Grammy awards?
          example: false
      required:
        - name
        - grammy
    AlbumResponseDto:
      type: object
      properties:
        id:
          type: string
          description: album ID (uuid v4)
          example: 51699c8f-26a0-493e-be62-8fe9d07c5a40
        name:
          type: string
          description: album name
          example: Cropping the Aftermath
        year:
          type: number
          minimum: 1
          description: album year
          example: 2020
        artistId:
          type: string
          nullable: true
          description: ID of the artist who created this album (or null)
          example: 48d89280-a918-4de6-8141-c34e9cf6283e
      required:
        - id
        - name
        - year
        - artistId
    CreateAlbumDto:
      type: object
      properties:
        name:
          type: string
          description: album name
          example: In the Days of Jupiter
        year:
          type: number
          minimum: 1
          description: album year
          example: 2010
        artistId:
          type: string
          nullable: true
          description: ID of the artist who created this album (or null)
          example: 2dd25d8d-21af-40b1-a273-1f06026fc618
      required:
        - name
        - year
        - artistId
    UpdateAlbumInfoDto:
      type: object
      properties:
        name:
          type: string
          description: album name
          example: Weathering
        year:
          type: number
          minimum: 1
          description: album year
          example: 2011
        artistId:
          type: string
          nullable: true
          description: ID of the artist who created this album (or null)
          example: 5f800c90-e1f0-483e-ab8d-1efff5f904e7
      required:
        - name
        - year
        - artistId
...